+++ cli/src/core/ast/deletions/delete_from_universal_sdk.rs
@@ -51,18 +51,10 @@ pub(crate) fn delete_from_universal_sdk<'a>(
 
     let import_source = format!("@{}/{}", kebab_app_name, kebab_case_name);
     delete_import_statement(allocator, app_program_ast, &import_source)?;
-    let serialized_import_source = format!("@{}/{}/serialized", kebab_app_name, kebab_case_name);
-    delete_import_statement(allocator, app_program_ast, &serialized_import_source)?;
 
     let blueprint_import_source =
         format!("@{}/{}{}", kebab_app_name, "blueprint-", kebab_case_name);
     delete_import_statement(allocator, app_program_ast, &blueprint_import_source)?;
-    let blueprint_serialized_import_source = format!("{}/serialized", blueprint_import_source);
-    delete_import_statement(
-        allocator,
-        app_program_ast,
-        &blueprint_serialized_import_source,
-    )?;
 
     delete_from_universal_sdk_function(allocator, app_program_ast, camel_case_name)?;
 
@@ -120,7 +112,7 @@ mod tests {
         import {
           BetterAuthConfig,
           IamSdkClient as IamBetterAuthSdkClient
-        } from '@forklaunch/blueprint-iam-better-auth/serialized';
+        } from '@forklaunch/blueprint-iam-better-auth';
         import { universalSdk, RegistryOptions } from '@forklaunch/universal-sdk';
         import { createAuthClient } from 'better-auth/client';
         import { inferAdditionalFields } from 'better-auth/client/plugins';