+++ framework/core/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @forklaunch/core
 
+## 0.15.12
+
+### Patch Changes
+
+- fix hyper express header, fix tests
+- Updated dependencies
+  - @forklaunch/validator@0.10.22
+  - @forklaunch/common@0.6.22
+
+## 0.15.11
+
+### Patch Changes
+
+- Update package versions, and add x-powered-by forklaunch
+- Updated dependencies
+  - @forklaunch/validator@0.10.21
+  - @forklaunch/common@0.6.21
+
+## 0.15.10
+
+### Patch Changes
+
+- update internal package versions
+- Updated dependencies
+  - @forklaunch/validator@0.10.20
+  - @forklaunch/common@0.6.20
+
+## 0.15.9
+
+### Patch Changes
+
+- Updates for openapi publishing mode
+
+## 0.15.8
+
+### Patch Changes
+
+- update package versions
+- Updated dependencies
+  - @forklaunch/validator@0.10.19
+  - @forklaunch/common@0.6.19
+
 ## 0.15.7
 
 ### Patch Changes